add new functions2

+++ application/models/vote.go
@@ -3,6 +3,6 @@ package models
 type Vote struct {
 	Nickname   string `json:"nickname"`
 	Voice      int32  `json:"voice"`
-	Thread     int64  `json:"thread,omitempty"`
+	Thread     int32  `json:"thread,omitempty"`
 	ThreadSlug string `json:"thread_slug,omitempty"`
 }
